conduit_outreach_pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "conduit-outreach-pipeline"
version = "1.1.0"
description = "Full outreach: ingest, ConduitScore (cached), A/B/C sequences, Jinja2+snippets, Google Sheets, Gmail send"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "jinja2>=3.1.0",
  "typer>=0.12.0",
  "rich>=13.0.0",
  "trafilatura>=1.8.0",
  "gspread>=6.0.0",
  "google-auth>=2.0.0",
  "google-api-python-client>=2.100.0",
  "google-auth-oauthlib>=1.0.0",
  "google-auth-httplib2>=0.2.0",
]

[project.optional-dependencies]
heavy = ["crawl4ai>=0.4.0", "scrapy>=2.11.0"]

[project.scripts]
conduit-op = "conduit_outreach_pipeline.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/conduit_outreach_pipeline"]
